refactor: reorganize project structure for clearer data & UI separation
- Moved model/ → data/remote/dto/ (DTOs for API)
- Moved network/ → data/remote/ (remote data sources)
- Moved layouts/ → ui/common/ (shared UI components)
- data/local/: local data storage (TokenManager, SharedPreferences)
- data/remote/: API and network logic
- ui/common/: reusable UI components and factories
