+++ app/src/main/java/com/example/active_portfolio_mobile/data/remote/RetrofitClient.kt
@@ -25,6 +25,8 @@ import okhttp3.MediaType.Companion.toMediaType
 object RetrofitClient {
     // Base URL for Android emulator to access local development backend
     private const val DEV_BASE_URL = "http://10.0.2.2:3000/"
+    private const val PROD_BASE_URL = "https://activeportfolio.onrender.com/"
+//    private const val DEV_BASE_URL = "http://10.0.2.2:1339/"
 
     private const val BASE_URL = DEV_BASE_URL
