+++ app/src/main/java/com/example/active_portfolio_mobile/data/remote/network/AdventureService.kt
@@ -1,6 +1,6 @@
 package com.example.active_portfolio_mobile.network
 
-import com.example.active_portfolio_mobile.model.Adventure
+import com.example.active_portfolio_mobile.data.remote.dto.Adventure
 import retrofit2.Response
 import retrofit2.http.GET
