+++ app/src/main/java/com/example/active_portfolio_mobile/ui/common/MainLayout.kt
@@ -16,14 +16,13 @@ import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import com.example.active_portfolio_mobile.Navigation.LocalNavController
+import com.example.active_portfolio_mobile.navigation.LocalNavController
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
-import androidx.compose.material.icons.filled.Call
 import androidx.compose.material.icons.filled.Home
 import androidx.compose.material.icons.filled.Person
 import androidx.compose.material3.BottomAppBar
-import com.example.active_portfolio_mobile.Navigation.Routes
+import com.example.active_portfolio_mobile.navigation.Routes
 
 /**
  * The main layout to be shared across screens.
